+++ examples/basic/assertion/ensure.php
@@ -12,6 +12,6 @@ function ensure_we_are_in_the_future(): void
     check(
         'Check if we are in the future',
         'We are not in the future 😱',
-        fn () => (!usleep(500_000) && new \DateTime() > new \DateTime('2015-10-21'))
+        static fn () => (!usleep(500_000) && new \DateTime() > new \DateTime('2015-10-21'))
     );
 }