+++ examples/basic/cache/simple.php
@@ -10,9 +10,9 @@ use function Castor\io;
 #[AsTask(description: 'Cache a simple call')]
 function simple(): void
 {
-    io()->writeln(cache('my-key', fn () => 'SALUT'));
+    io()->writeln(cache('my-key', static fn () => 'SALUT'));
     // Should returns the same things
-    io()->writeln(cache('my-key', fn () => 'HELLO'));
+    io()->writeln(cache('my-key', static fn () => 'HELLO'));
     // Should returns the recomputed value
-    io()->writeln(cache('my-key', fn () => 'HOLA', true));
+    io()->writeln(cache('my-key', static fn () => 'HOLA', true));
 }